restructured galaxy and sed
