+++ examples/parametric/generate_sed.py
@@ -5,10 +5,9 @@ import matplotlib.pyplot as plt
 from synthesizer.filters import SVOFilterCollection
 from synthesizer.grid import Grid
 from synthesizer.parametric.sfzh import SFH, ZH, generate_sfzh
-from synthesizer.parametric.galaxy import Galaxy
+from synthesizer.galaxy.parametric import ParametricGalaxy as Galaxy
 from synthesizer.plt import single, single_histxy, mlabel
 from unyt import yr, Myr
-from synthesizer.igm import Madau96, Inoue14
 from astropy.cosmology import Planck18 as cosmo
