+++ synthesizer/sed/data/LAFcoeff.txt
