+++ tests/test_igm.py
@@ -8,8 +8,10 @@ from synthesizer.igm import Inoue14, Madau96
 def i14():
     return Inoue14()
 
+
 def test_I14_name(i14):
     assert type(i14.name) is str
+
 
 def test_I14_transmission(i14):
     lam = np.loadtxt('tests/test_sed/lam.txt')
